refactor(mix): move DiscoveryDownloadWorker from :data:download to :core:data
Original placement at :data:download was wrong - :core:data depends
on :data:download (via its existing dependency direction would be a
cycle), so StashDiscoveryWorker can't chain into a worker that lives
in :data:download. The worker's imports already only reference
com.stash.core.data.* and com.stash.core.model.* symbols, so this
is purely a file-location fix.

Now sits alongside TrackDownloadWorker (the analog it duplicates
per-track flow from) under core/data/.../sync/workers/.
