+++ core/data/src/main/kotlin/com/stash/core/data/sync/workers/DiscoveryDownloadWorker.kt
@@ -1,4 +1,4 @@
-package com.stash.data.download
+package com.stash.core.data.sync.workers
 
 import android.content.Context
 import android.content.pm.ServiceInfo
